pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tcpe"
version = "0.1.0"
description = "Transportable Causal Perturbation Engine (Phase 2 scaffold)"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { text = "MIT" }
authors = [{ name = "TCPE Contributors" }]
dependencies = [
  "anndata>=0.10,<0.12",
  "numpy>=1.26,<2.0",
  "pandas>=2.2,<3.0",
  "pydantic>=2.10,<3.0",
  "pot>=0.9,<1.0",
  "pyyaml>=6.0,<7.0",
  "scanpy>=1.10,<2.0",
  "scipy>=1.13,<2.0",
  "torch>=2.5,<3.0",
  "wandb>=0.18,<1.0",
]

[project.scripts]
tcpe = "tcpe.cli:main"

[project.optional-dependencies]
dev = [
  "mypy>=1.14,<2.0",
  "pre-commit>=4.0,<5.0",
  "pytest>=8.3,<9.0",
  "pytest-cov>=6.0,<7.0",
  "ruff>=0.8,<1.0",
  "types-PyYAML>=6.0,<7.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.11"
files = ["src"]
mypy_path = "src"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
strict_optional = true
check_untyped_defs = true
ignore_missing_imports = true
